Timeseries uses Aggregate API endpoint.
The pagesize set from outside of the widget controls the number of entities on
the visualization. I didn't simply pass that to the Aggregator because it
defines the TOTAL number of elements (i.e. if there's only one entity with 200
entries, it'll be the only shown). I think controlling the number of entities
makes more sense.

The bad side is that I'm not limiting the actual Aggregator's pagesize. Even
though we'll only show 10 entities (for example), we'll get them all. This
causes a big performance hit on huge datasets. This definitely is something
worth looking into afterwards.